+++ src/main/java/com/buuz135/togetherforever/command/TeamLeaveCommand.java
@@ -1,5 +1,6 @@
 package com.buuz135.togetherforever.command;
 
+import com.buuz135.togetherforever.api.IPlayerInformation;
 import com.buuz135.togetherforever.api.ITogetherTeam;
 import com.buuz135.togetherforever.api.TogetherForeverAPI;
 import com.buuz135.togetherforever.api.command.SubCommandAction;
@@ -7,6 +8,7 @@ import com.buuz135.togetherforever.api.data.DefaultPlayerInformation;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.command.PlayerNotFoundException;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
@@ -25,6 +27,12 @@ public class TeamLeaveCommand extends SubCommandAction {
             if (togetherTeam != null) {
                 TogetherForeverAPI.getInstance().removePlayerFromTeam(togetherTeam, DefaultPlayerInformation.createInformation(CommandBase.getCommandSenderAsPlayer(sender)));
                 CommandBase.getCommandSenderAsPlayer(sender).sendMessage(new TextComponentString(TextFormatting.GREEN + "You successfully left your team."));
+                for (IPlayerInformation playerInformation : togetherTeam.getPlayers()) {
+                    EntityPlayerMP playerMP = playerInformation.getPlayer();
+                    if (playerMP != null) {
+                        playerMP.sendMessage(new TextComponentString(TextFormatting.RED + "Player " + sender.getName() + " left your team."));
+                    }
+                }
                 return true;
             }
         } catch (PlayerNotFoundException e) {