+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Added more safety checks for player info comparison 
 + Added a config option to delay the team syncing action
 + Fixed team owners no being able to kick team members
++ Added more descriptive leave message for the other members of the team
 
 # 1.0.1
 + Safe check for missing players when accepting invites